Implement RoleService and UserService (to be tested)

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/SimpleRoleService.java
@@ -2,14 +2,20 @@ package at.ac.tuwien.sepm.groupphase.backend.service.impl;
 
 import at.ac.tuwien.sepm.groupphase.backend.entity.Role;
 import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundException2;
+import at.ac.tuwien.sepm.groupphase.backend.exception.ValidationException;
 import at.ac.tuwien.sepm.groupphase.backend.repository.RoleRepository;
 import at.ac.tuwien.sepm.groupphase.backend.service.RoleService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.lang.invoke.MethodHandles;
+import java.util.List;
+import java.util.Locale;
+import java.util.Optional;
 
+@Service
 public class SimpleRoleService implements RoleService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
@@ -21,21 +27,22 @@ public class SimpleRoleService implements RoleService {
     }
 
     @Override
-    public Role create(Role role) {
-        Role newRole = new Role(null, role.getName(), null);
+    public Role create(Role role) throws ValidationException {
+        Role newRole = new Role(null, role.getName().toUpperCase(Locale.ROOT), role.getApplicationUsers());
+        if (false) {
+            // ToDo: sanity check name (only A-Za-z0-9)
+            // ToDo: sanity check users (only existing users)
+            throw new ValidationException("Role invalid");
+        }
+
         return this.roleRepository.save(newRole);
     }
 
     @Override
     public Role findRoleByName(String name) throws NotFoundException2 {
-        if (this.roleExistsByName(name)) {
-            Role role = this.roleRepository.findRoleByName(name);
-            if (role == null) {
-                // should only happen if between the execution of first and third line the role is deleted
-                throw new NotFoundException2("Role not found.");
-            } else {
-                return role;
-            }
+        Optional<Role> role = this.roleRepository.findRoleByName(name);
+        if (role.isPresent()) {
+            return role.get();
         } else {
             throw new NotFoundException2("Role not found.");
         }
@@ -43,22 +50,42 @@ public class SimpleRoleService implements RoleService {
 
     @Override
     public boolean roleExistsByName(String name) {
-        Role role = roleRepository.findRoleByName(name);
-        if (role == null) {
-            return false;
-        } else {
-            return true;
-        }
+        Optional<Role> role = roleRepository.findRoleByName(name);
+        return role.isPresent();
+    }
+
+    @Override
+    public boolean roleExistsById(Long id) {
+        Optional<Role> role = roleRepository.findRoleById(id);
+        return role.isPresent();
     }
 
     @Override
     public void deleteById(Long id) throws NotFoundException2 {
-        this.roleRepository.deleteById(id);
+        if (roleExistsById(id)) {
+            this.roleRepository.deleteById(id);
+        } else {
+            throw new NotFoundException2("Role cannot be found.");
+        }
     }
 
     @Override
     public void deleteByName(String name) throws NotFoundException2 {
-        Role role = this.findRoleByName(name);
-        this.deleteById(role.getId());
+        if (roleExistsByName(name)) {
+            try {
+                Role role = this.findRoleByName(name);
+                this.deleteById(role.getId());
+            } catch (NotFoundException2 exc) {
+                throw new NotFoundException2("Role cannot be found.", exc);
+            }
+        } else {
+            throw new NotFoundException2("Role cannot be found.");
+        }
+    }
+
+    @Override
+    public List<Role> findAll() {
+        LOGGER.debug("Find all roles");
+        return roleRepository.findAll();
     }
 }
